+++ tests/test_models.py
@@ -1,5 +1,3 @@
-import warnings
-
 import pytest
 
 import guitarpro as gp
@@ -24,14 +22,6 @@ def test_duration(value, isDotted, tuplet):
     new_dur = gp.Duration.fromTime(time)
     assert isinstance(new_dur.value, int)
     assert time == new_dur.time
-
-
-def test_double_dot_warning(recwarn):
-    warnings.simplefilter('always')
-    d = gp.Duration()
-    assert len(recwarn) == 0
-    with pytest.deprecated_call():
-        d.isDoubleDotted = True
 
 
 def test_beat_start_in_measure():